+++ .github/workflows/test.yaml
@@ -22,11 +22,11 @@ jobs:
         run: go mod tidy
       - name: Test
         run: go test -v ./...
-      - name: Run Demo
-        run: go run main.go
-      - name: Upload Test Artifact
-        uses: actions/upload-artifact@v2
-        with:
-          name: proto-demo
-          path: proto-demo
-          if-no-files-found: error
+#      - name: Run Demo
+#        run: go run main.go
+#      - name: Upload Test Artifact
+#        uses: actions/upload-artifact@v2
+#        with:
+#          name: proto-demo
+#          path: proto-demo
+#          if-no-files-found: error
