+++ pkg/parse_test.go
@@ -21,7 +21,7 @@ func TestGetKeys(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			d, _ := parse.Parse(tt.name, tt.args, "{{", "}}")
-			got := GetParseKeys(d[tt.name].Root)
+			got := getParseKeys(d[tt.name].Root)
 			if !reflect.DeepEqual(got, tt.want) {
 				t.Errorf("GetParseKeys() = %v, want %v", got, tt.want)
 			}
